requirements.txt
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
scipy>=1.11.0
